Update : - databse - chức năng đồng bộ - chức năng thanh toán online/offline - chức năng tìm kiếm online trong thanh toán Error phát sinh: - Định danh thẻ --> do update database - Report --> do update database

+++ app/src/main/java/views/ecpay/com/postabletecpay/util/entities/request/EntityBillingOnline/BodyBillingOnlineRequest.java
@@ -27,7 +27,7 @@ public class BodyBillingOnlineRequest implements Serializable
     private String session;
     @SerializedName("bill-id")
     @Expose
-    private Integer billId;
+    private Long billId;
     @SerializedName("amount")
     @Expose
     private Long amount;
@@ -90,11 +90,11 @@ public class BodyBillingOnlineRequest implements Serializable
         this.session = session;
     }
 
-    public Integer getBillId() {
+    public Long getBillId() {
         return billId;
     }
 
-    public void setBillId(Integer billId) {
+    public void setBillId(Long billId) {
         this.billId = billId;
     }
